+++ src/core/dcx-server.ts
@@ -1,3 +1,4 @@
+import { Web5PlatformAgent } from '@web5/agent';
 import { Web5 } from '@web5/api';
 import { generateMnemonic } from 'bip39';
 import crypto from 'crypto';
@@ -10,9 +11,7 @@ import { DcxServerError } from '../utils/error.js';
 import { readFileToJSON, readFileToString } from '../utils/file-system.js';
 import { stringifier } from '../utils/json.js';
 import { Time } from '../utils/time.js';
-import { DidManager } from './did-manager.js';
-import { DwnManager } from './dwn-manager.js';
-import { Web5PlatformAgent } from '@web5/agent';
+import { Web5Manager } from './web5-manager.js';
 
 type DcxServerOptions = { manifests?: CredentialManifest[] };
 
@@ -23,11 +22,11 @@ export class DcxServer extends DcxEnv {
   constructor(options: DcxServerOptions) {
     super();
     this.isInitialized = !!this.WEB5_CONNECT_RECOVERY_PHRASE;
-    this.manifests = DwnManager.manifests = options.manifests ?? [];
+    this.manifests = Web5Manager.manifests = options.manifests ?? [];
   }
 
   public useManifest(name: string, manifest: CredentialManifest): void {
-    DwnManager.manifests.push({ ...manifest, name });
+    Web5Manager.manifests.push({ ...manifest, name });
   }
 
   /**
@@ -96,17 +95,17 @@ export class DcxServer extends DcxEnv {
         techPreview: { dwnEndpoints: DcxEnv.DWN_ENDPOINTS },
       });
 
-      DwnManager.web5 = web5;
-      DwnManager.agent = web5.agent as Web5PlatformAgent;
+      Web5Manager.web5 = web5;
+      Web5Manager.agent = web5.agent as Web5PlatformAgent;
 
-      const { did: bearerDid } = await DwnManager.agent.identity.get({ didUri: did }) ?? {};
+      const { did: bearerDid } = await Web5Manager.agent.identity.get({ didUri: did }) ?? {};
       if (!bearerDid) {
         throw new DcxServerError('Failed to get bearer DID');
       }
 
-      DidManager.did = did;
-      DidManager.bearerDid = bearerDid;
-      DidManager.portableDid = await bearerDid.export();
+      Web5Manager.did = did;
+      Web5Manager.bearerDid = bearerDid;
+      Web5Manager.portableDid = await bearerDid.export();
 
       if (!this.WEB5_CONNECT_RECOVERY_PHRASE && newSeedPhrase) {
         console.info(
@@ -121,13 +120,13 @@ export class DcxServer extends DcxEnv {
           'to reuse the DID created for this DCX server, set ' +
           'WEB5_CONNECT_RECOVERY_PHRASE and WEB5_CONNECT_PASSWORD in .env',
         );
-        await writeFile('portable.json', stringifier(DidManager.portableDid));
+        await writeFile('portable.json', stringifier(Web5Manager.portableDid));
       }
 
       console.log('Web5 connection initialized!');
       // console.log('this  =>', this, '\n~~~~~~~~~~~~~~~~~~~~~~\n');
-      // console.log('web5  =>', DwnManager.web5, '\n~~~~~~~~~~~~~~~~~~~~~~\n');
-      // console.log('did   =>', DidManager.bearerDid, '\n~~~~~~~~~~~~~~~~~~~~~~\n');;
+      // console.log('web5  =>', Web5Manager.web5, '\n~~~~~~~~~~~~~~~~~~~~~~\n');
+      // console.log('did   =>', Web5Manager.bearerDid, '\n~~~~~~~~~~~~~~~~~~~~~~\n');;
       this.isInitialized = true;
     } catch (error: any) {
       console.error('DcxServer.setup', error);
@@ -147,8 +146,8 @@ export class DcxServer extends DcxEnv {
       console.log('DCX server running!');
 
       while (true) {
-        const { records = [], cursor: nextCursor } = await DwnManager.web5.dwn.records.query({
-          from: DidManager.bearerDid.uri,
+        const { records = [], cursor: nextCursor } = await Web5Manager.web5.dwn.records.query({
+          from: Web5Manager.bearerDid.uri,
           message: {
             filter: {
               protocol: credentialIssuerProtocol.protocol,
@@ -166,8 +165,8 @@ export class DcxServer extends DcxEnv {
 
         const recordReads = await Promise.all(
           records.map(async (recordResponse: { id: any }) => {
-            const { record } = await DwnManager.web5.dwn.records.read({
-              from: DidManager.bearerDid.uri,
+            const { record } = await Web5Manager.web5.dwn.records.read({
+              from: Web5Manager.bearerDid.uri,
               message: {
                 filter: {
                   recordId: recordResponse.id,
@@ -181,7 +180,7 @@ export class DcxServer extends DcxEnv {
         for (const record of recordReads) {
           if (record.id != lastRecordId) {
             if (record.protocolPath === 'application') {
-              const applicationManifest = DwnManager.manifests.find(
+              const applicationManifest = Web5Manager.manifests.find(
                 (manifest: CredentialManifest) =>
                   manifest.presentation_definition.id === record.schema,
               );
@@ -228,7 +227,7 @@ export class DcxServer extends DcxEnv {
     }
 
     try {
-      await DwnManager.setup();
+      await Web5Manager.setup();
     } catch (error) {
       console.error('Failed to setup DCX DWN', error);
     }
